requirements/base.txt
aiohttp
mautrix-appservice
ruamel.yaml
python-magic
SQLAlchemy
alembic
Markdown
future-fstrings
telethon-session-sqlalchemy
